restructure style and subcomponent containters
