+++ src/components/ReusableComponents/ButtonComponent/button.css
@@ -1,0 +1,37 @@
+/* BUTTON COMPONENT */
+.button_box {
+    margin: 16px;
+    
+    button {
+        width: 120px;
+        height: 50px;
+        padding: 8px 12px;
+        
+        &:hover {filter: brightness(1.2);}
+    }
+}
+.button_default {
+    background-color: var(--blue);
+    color: var(--white);
+    border: 2px solid var(--red);
+}
+
+.button_red {
+    background-color: var(--red);
+    color: var(--white);
+    border: 2px solid var(--blue);
+}
+
+.button_white {
+    background-color: var(--white);
+    color: var(--blue);
+    border: 2px solid var(--red);
+}
+
+.button_blue {
+    background-color: var(--blue);
+    color: var(--white);
+    border: 2px solid var(--red);
+}
+
+/* END BUTTON */
