+++ src/views/Home.jsx
@@ -1,0 +1,12 @@
+// HOME VIEW PAGE
+
+import HomeHero from "../components/ReusableComponents/HeroComponent/Hero";
+
+export default function Home() {
+    
+    return(
+        <main>
+            <HomeHero Title1="Senior Mobile" Title2="Tax Services" SubTitle="Don't Drive... Make The Call" />
+        </main>
+    );
+};
