--- src/views/Index.jsx
@@ -1,11 +1,0 @@
-// HOME VIEW PAGE (Index)
-
-export default function Index() {
-    
-    return(
-        <main>
-            <h1>TEST Home Page</h1>
-            <h3>This is a test...</h3>
-        </main>
-    );
-};
