+++ src/views/Services.jsx
@@ -1,7 +1,7 @@
 // SERVICES VIEW PAGE
 
 // IMPORT SUBCOMPONENTS
-import Button from "../components/ReusableComponents/Button";
+import Button from "../components/ReusableComponents/ButtonComponent/Button";
 
 export default function Services() {
